gh-1344 Sharing model between extension and main app
