+++ Multisig.xcodeproj/xcshareddata/xcschemes/MultisigIntegrationTests.xcscheme
@@ -7,7 +7,7 @@
       buildImplicitDependencies = "YES">
    </BuildAction>
    <TestAction
-      buildConfiguration = "Debug.Production"
+      buildConfiguration = "Debug.Development.Rinkeby"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
@@ -31,7 +31,7 @@
       </Testables>
    </TestAction>
    <LaunchAction
-      buildConfiguration = "Debug.Production"
+      buildConfiguration = "Debug.Development.Rinkeby"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       launchStyle = "0"
@@ -51,17 +51,17 @@
       </MacroExpansion>
    </LaunchAction>
    <ProfileAction
-      buildConfiguration = "Debug.Production"
+      buildConfiguration = "Release.Staging.Rinkeby"
       shouldUseLaunchSchemeArgsEnv = "YES"
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
    </ProfileAction>
    <AnalyzeAction
-      buildConfiguration = "Debug.Production">
+      buildConfiguration = "Debug.Staging.Rinkeby">
    </AnalyzeAction>
    <ArchiveAction
-      buildConfiguration = "Debug.Production"
+      buildConfiguration = "Release.Staging.Rinkeby"
       revealArchiveInOrganizer = "YES">
    </ArchiveAction>
 </Scheme>
