+++ Multisig.xcodeproj/xcshareddata/xcschemes/NotificationServiceExtension.xcscheme
@@ -38,7 +38,7 @@
       </BuildActionEntries>
    </BuildAction>
    <TestAction
-      buildConfiguration = "Debug.Production"
+      buildConfiguration = "Debug.Development.Rinkeby"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
@@ -46,7 +46,7 @@
       </Testables>
    </TestAction>
    <LaunchAction
-      buildConfiguration = "Debug.Production"
+      buildConfiguration = "Debug.Development.Rinkeby"
       selectedDebuggerIdentifier = ""
       selectedLauncherIdentifier = "Xcode.IDEFoundation.Launcher.PosixSpawn"
       launchStyle = "0"
@@ -69,7 +69,7 @@
       </BuildableProductRunnable>
    </LaunchAction>
    <ProfileAction
-      buildConfiguration = "Debug.Production"
+      buildConfiguration = "Release.Development.Rinkeby"
       shouldUseLaunchSchemeArgsEnv = "YES"
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
@@ -87,10 +87,10 @@
       </BuildableProductRunnable>
    </ProfileAction>
    <AnalyzeAction
-      buildConfiguration = "Debug.Production">
+      buildConfiguration = "Debug.Development.Rinkeby">
    </AnalyzeAction>
    <ArchiveAction
-      buildConfiguration = "Debug.Production"
+      buildConfiguration = "Release.Development.Rinkeby"
       revealArchiveInOrganizer = "YES">
    </ArchiveAction>
 </Scheme>
