+++ Shared/CharacterSet+Hex.swift
