+++ Shared/CoreDataProtocol.swift
