+++ Shared/Safe+CoreDataClass.swift
