+++ Shared/Safe+CoreDataProperties.swift
